20250821_정우철_신우 복사(라우팅 관련)

+++ src/main/java/mes/domain/entity/RoutingProc.java
@@ -23,7 +23,7 @@ public class RoutingProc extends AbstractAuditModel{
 	int id;
 	
 	@Column(name = "\"ProcessOrder\"")
-	Float processOrder;
+	Integer processOrder;
 	
 	@Column(name = "\"StandardTime\"")
 	Float standardTime;
@@ -32,7 +32,7 @@ public class RoutingProc extends AbstractAuditModel{
 	String description;
 	
 	@Column(name = "\"Process_id\"")
-	Float processId;
+	Integer processId;
 	
 	@Column(name = "\"Routing_id\"")
 	Integer routingId;
